fix test case for session renew
